"alertas de creación, editado y borrado"

+++ src/app/components/product-list/product-list.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Product } from '../../common/product';
 import { ProductService } from '../../services/product.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-product-list',
@@ -26,12 +27,29 @@ export class ProductListComponent implements OnInit{
     );
   }
   deleteProductById(id: number) {
-    this.productService.deleteProduct(id).subscribe(
-      data => {
-        console.log('Product deleted');
-        this.listProducts();
+    Swal.fire({
+      title: "¿Está seguro?",
+      text: "No puede revertir los cambios!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Eliminar!",
+      cancelButtonText: "Cancelar"
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.productService.deleteProduct(id).subscribe(
+          () => this.listProducts()
+          
+        );
+        Swal.fire({
+          title: "Producto borrado!",
+          text: "Producto eliminado correctamente.",
+          icon: "success"
+        });
       }
-    );
+    });
+    
   }
